fix: resolve all clippy warnings (88 warnings → 0)
Auto-fixed (cargo clippy --fix):
- Redundant closures, format! in println, map_or→is_none_or
- Collapsible if/else, borrow deref, redundant imports
- Compound assignments, is_multiple_of, push_str, range contains

Manual fixes:
- .max().min() → .clamp() (countdown models, render)
- &PathBuf → &Path (lifecycle.rs)
- Manual counter → .enumerate() (service.rs)
- to_rrule_freq/to_rrule_day/to_theme → as_* (clippy::wrong_self_convention)
- #[derive(Default)] on ICalendarService unit struct
- #[allow(dead_code)] on WIP resize module and unused fields/methods
- #[allow(clippy::too_many_arguments)] on 10 functions (8-15 params)

+++ src/ui_egui/app/navigation.rs
@@ -41,7 +41,7 @@ impl CalendarApp {
 }
 
 fn shift_month_preserving_day(current: NaiveDate, delta_months: i32) -> NaiveDate {
-    let total_months = (current.year() * 12) as i32 + (current.month() as i32 - 1) + delta_months;
+    let total_months = (current.year() * 12) + (current.month() as i32 - 1) + delta_months;
     let new_year = total_months.div_euclid(12);
     let new_month = total_months.rem_euclid(12) + 1;
     clamp_day(new_year, new_month as u32, current.day())
